Recreate toggle button 🌚🌝
